Allow passing through anonymous match groups like {}.
Ideally these would be non-matching groups but I don't want to change the
str_to_re() function. re dunderstandably doesn't like named match groups with
empty names, nor does't it lke multiple named match groups with the same name.
So we have to think up a unique name for each anonymous match group. I pulled
in random for this and went "junkXXXX" but we could just as well have kept a
counter around somewhere.

Also changed the value checks on the string.Formater.pare() iterator because
normally field_name is None but in this case it is str(''). So no we are
matching on None, not falsy. Except for conversion because that is apparantly
a real but empty thing whenever field_name is specified.

Also made it so any None values are stripped away from ret before
saving/printing tags so that I could remove the the junk tags before hand. I
let empty values pass through so that maybe we could use them to unset tags?
How does that work anyway?