Add reference to zlib license.
Couldn't be bothered doubling the number of files in the repo just for this.
Mayb when I write a readme I will. For now just look at the usage message near
the top of the file.

Also, do I really have to include my name in a copyright claimer? Lame.