Minor fixes for eyeD3 API changes
Also a fix for an error thrown when fetching an empty comment.

Not sure when the track setting API changed but it looks like it just
takes a tuple now and in 0.6 so this should be fine.

Also looks like we should be using eyed3.load() or something instead of
Tag(), but whatever.